+++ extensions/pins/src/index.tsx
@@ -7,6 +7,7 @@ import {
   Clipboard,
   environment,
   getPreferenceValues,
+  getSelectedText,
   Icon,
   launchCommand,
   LaunchType,
@@ -20,16 +21,16 @@ import { useCachedState } from "@raycast/utils";
 import OpenAllMenuItem from "./components/OpenAllMenuItem";
 import PinMenuItem from "./components/PinMenuItem";
 import RecentApplicationsList from "./components/RecentApplicationsList";
-import { SupportedBrowsers } from "./lib/browser-utils";
 import { KEYBOARD_SHORTCUT, StorageKey } from "./lib/constants";
 import { createNewGroup, Group, useGroups } from "./lib/Groups";
 import { getGroupIcon } from "./lib/icons";
 import { useLocalData } from "./lib/LocalData";
 import { copyPinData, createNewPin, Pin, sortPins, usePins } from "./lib/Pins";
-import { Placeholders } from "./lib/placeholders";
 import { cutoff, getStorage, setStorage } from "./lib/utils";
-import { ExtensionPreferences } from "./lib/preferences";
+import { ExtensionPreferences, GroupDisplaySetting } from "./lib/preferences";
 import { PinsMenubarPreferences } from "./lib/preferences";
+import PinsPlaceholders from "./lib/placeholders";
+import { utils } from "placeholders-toolkit";
 
 /**
  * Raycast menu bar command providing quick access to pins.
@@ -70,19 +71,16 @@ export default function ShowPinsCommand() {
           const inapplicablePins = [];
           for (const pin of pins) {
             const targetRaw = pin.url.startsWith("~") ? pin.url.replace("~", os.homedir()) : pin.url;
-            const placeholders = Placeholders.allPlaceholders;
+            const placeholders = PinsPlaceholders;
             let containsPlaceholder = false;
             let passesTests = true;
             let ruleCount = 0;
-            for (const [placeholderText, placeholderValue] of Object.entries(placeholders)) {
-              if (
-                targetRaw.includes(placeholderText) ||
-                placeholderValue.aliases?.some((alias) => targetRaw.includes(alias))
-              ) {
+            for (const placeholder of placeholders) {
+              if (targetRaw.match(placeholder.regex)) {
                 containsPlaceholder = true;
-                for (const rule of placeholderValue.rules) {
+                for (const rule of placeholder.rules || []) {
                   ruleCount++;
-                  if (!(await rule(targetRaw, localData))) {
+                  if (!(await rule(targetRaw, localData as unknown as { [key: string]: unknown }))) {
                     passesTests = false;
                   }
                 }
@@ -123,6 +121,28 @@ export default function ShowPinsCommand() {
     if (memberPins.length == 0 && subgroupPins.length == 0) {
       return null;
     }
+
+    if (preferences.groupDisplaySetting === GroupDisplaySetting.Subsections) {
+      return (
+        <MenuBarExtra.Section title={group.name} key={group.id}>
+          {memberPins.map((pin) => (
+            <PinMenuItem
+              pin={pin}
+              relevant={relevantPins.find((p) => p.id == pin.id) != undefined && !preferences.showInapplicablePins}
+              preferences={preferences}
+              localData={localData}
+              setPins={setPins}
+              key={pin.id}
+            />
+          ))}
+          {children.map((g) => getSubsections(g, groups))}
+          {subgroupPins.length > 0 ? (
+            <OpenAllMenuItem pins={allPins.filter((p) => p.group == group.name)} submenuName={group.name} />
+          ) : null}
+        </MenuBarExtra.Section>
+      );
+    }
+
     return (
       <MenuBarExtra.Submenu
         title={
@@ -157,10 +177,13 @@ export default function ShowPinsCommand() {
     );
   };
 
-  const groupSubmenus = groups
-    .filter((g) => g.parent == undefined)
-    .map((group) => getSubsections(group, groups))
-    .filter((g) => g != null);
+  const groupSubmenus =
+    preferences.groupDisplaySetting === GroupDisplaySetting.None
+      ? []
+      : groups
+          .filter((g) => g.parent == undefined)
+          .map((group) => getSubsections(group, groups))
+          .filter((g) => g != null);
 
   // Display the menu
   return (
@@ -170,7 +193,7 @@ export default function ShowPinsCommand() {
           <MenuBarExtra.Section title={preferences.showCategories ? "Pins" : undefined} key="pins">
             {allPins.length == 0 ? <MenuBarExtra.Item title="No pins yet!" /> : null}
             {allPins
-              .filter((p) => p.group == "None")
+              .filter((p) => (preferences.groupDisplaySetting !== GroupDisplaySetting.None ? p.group == "None" : true))
               .map((pin: Pin) => (
                 <PinMenuItem
                   pin={pin}
@@ -211,33 +234,38 @@ export default function ShowPinsCommand() {
                     false,
                     undefined,
                     undefined,
+                    [],
+                    "",
                   );
                 }}
               />
             ) : null}
-            {localData.selectedText.trim().length > 0 ? (
-              <MenuBarExtra.Item
-                title={`Pin Selected Text (${cutoff(localData.selectedText, 20)})`}
-                icon={Icon.Text}
-                tooltip="Pin the currently selected text as a text fragment"
-                shortcut={KEYBOARD_SHORTCUT.PIN_SELECTED_TEXT}
-                onAction={async () => {
-                  await createNewPin(
-                    localData.selectedText.substring(0, 50).trim(),
-                    localData.selectedText,
-                    "text-16",
-                    "None",
-                    "None",
-                    undefined,
-                    undefined,
-                    true,
-                    undefined,
-                    undefined,
-                  );
-                }}
-              />
-            ) : null}
-            {SupportedBrowsers.includes(localData.currentApplication.name) ? (
+
+            <MenuBarExtra.Item
+              title={`Pin Selected Text`}
+              icon={Icon.Text}
+              tooltip="Pin the currently selected text as a text fragment"
+              shortcut={KEYBOARD_SHORTCUT.PIN_SELECTED_TEXT}
+              onAction={async () => {
+                const text = await getSelectedText();
+                await createNewPin(
+                  text.substring(0, 50).trim(),
+                  text,
+                  "text-16",
+                  "None",
+                  "None",
+                  undefined,
+                  undefined,
+                  true,
+                  undefined,
+                  undefined,
+                  [],
+                  "",
+                );
+              }}
+            />
+
+            {utils.SupportedBrowsers.find((b) => b.name == localData.currentApplication.name) ? (
               <MenuBarExtra.Item
                 title={`Pin This Tab (${cutoff(localData.currentTab.name, 20)})`}
                 icon={Icon.AppWindow}
@@ -255,11 +283,14 @@ export default function ShowPinsCommand() {
                     false,
                     undefined,
                     undefined,
+                    [],
+                    "",
                   );
                 }}
               />
             ) : null}
-            {SupportedBrowsers.includes(localData.currentApplication.name) && localData.tabs.length > 1 ? (
+            {utils.SupportedBrowsers.find((b) => b.name == localData.currentApplication.name) &&
+            localData.tabs.length > 1 ? (
               <MenuBarExtra.Item
                 title={`Pin All Tabs (${localData.tabs.length})`}
                 icon={Icon.AppWindowGrid3x3}
@@ -288,6 +319,8 @@ export default function ShowPinsCommand() {
                       false,
                       undefined,
                       undefined,
+                      [],
+                      "",
                     );
                   }
                 }}
@@ -316,6 +349,8 @@ export default function ShowPinsCommand() {
                       false,
                       undefined,
                       undefined,
+                      [],
+                      "",
                     );
                   } else {
                     let newGroupName = "New File Group";
@@ -340,6 +375,8 @@ export default function ShowPinsCommand() {
                         false,
                         undefined,
                         undefined,
+                        [],
+                        "",
                       );
                     }
                   }
@@ -364,6 +401,8 @@ export default function ShowPinsCommand() {
                     false,
                     undefined,
                     undefined,
+                    [],
+                    "",
                   );
                 }}
               />
@@ -395,6 +434,8 @@ export default function ShowPinsCommand() {
                     false,
                     undefined,
                     undefined,
+                    [],
+                    "",
                   );
                 }}
               />
@@ -423,6 +464,8 @@ export default function ShowPinsCommand() {
                       false,
                       undefined,
                       undefined,
+                      [],
+                      "",
                     );
                   } else {
                     let newGroupName = "New Note Group";
@@ -445,6 +488,8 @@ export default function ShowPinsCommand() {
                         false,
                         undefined,
                         undefined,
+                        [],
+                        "",
                       );
                     }
                   }