+++ extensions/tana/src/api/helpers.ts
@@ -1,18 +1,22 @@
 import { getPreferenceValues } from "@raycast/api";
-import axios from "axios";
+import { TanaAPIHelper } from "./TanaAPIClient";
 
 export const TANA_URL = "https://europe-west1-tagr-prod.cloudfunctions.net";
 
 type Preferences = {
-  nodeApiToken: string;
+  workspaceApiToken: string;
 };
 export const prefs = getPreferenceValues<Preferences>();
 
-export type TanaResponse = string | { err: string };
-
-export const nodeApi = axios.create({
-  baseURL: TANA_URL,
-  headers: {
-    authorization: `Bearer ${prefs.nodeApiToken}`,
-  },
-});
+export const createPlainNode = async (name: string): Promise<string> => {
+  const helper = new TanaAPIHelper(prefs.workspaceApiToken);
+  const tananode = await helper.createNode(
+    {
+      name,
+    },
+    // 'SCHEMA'
+    // 'LIBRARY'
+    "INBOX",
+  );
+  return tananode.name;
+};