+++ extensions/browser-bookmarks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Browser Bookmarks Changelog
 
+## [Bug Fixes] - 2023-12-29
+
+- Fixed an issue where the selected item would not be the first one after the search text changed
+
+## [Support for Sidekick] - 2023-12-18
+
+- Add support for `Sidekick` browser
+
 ## [Support for Brave Nightly] - 2023-12-09
 
 - Add support for `Brave Nightly` browser.